Add a porcelain URL for `speedsolving.com`.
This is analogous to "git porcelain" — an API/URL that we can keep
supporting even if we change other URLs.

This is under the `@cubing/porcelain` namespace because that is a
hypothetical `npm` package name that we can publish in the future if we
want (but no one else can).